--- Bullet.java
@@ -1,10 +1,0 @@
-package Game;
-
-public class Bullet extends GameObject {
-	
-	Bullet(){
-		
-	}
-	
-	
-}
