--- Listener.java
@@ -1,5 +1,0 @@
-package Game;
-
-public class Listener {
-
-}
