--- main.java
@@ -1,11 +1,0 @@
-package Game;
-
-public class main {
-
-	public static void main(String[] args) {
-		Spaceship fighter = new Spaceship();
-		fighter.setY(20);
-		System.out.print(fighter.getX());
-	}
-
-}
